Custom live boot and install image
Modifications to initramfs image included in the true-image-touch live (.hdd) image.
Syslinux (removable drive) boot options are now:
           "live" image boot, "install", drop to "shell" within initramfs, and "boot" from harddrive.

Changes to the install.sh script were made to partition the install target
drive with boot, swap, config, and data partitions.

The kernel image, rootfs and initrd images are written to the target drive boot
partition, and grub is installed on this partition (probably /dev/sda1.)

From the installed partition, grub boots with an initial ramdisk which mounts the rootfs image
in loop-back and then switches root to the loopback mount.

This structure allows upgrading the root file system by writing a new rootfs image
file, instead of writing to the raw block device.  This allows an update in an atomic
operation (write a new rootfs image, then rename it to the active boot image) avoiding
corrupt upgrades in case of a power-failure during update.

	new file:   images/true-image-initramfs.bb       - new ramdisk image
	modified:   images/true-image-touch.bb           - use the new ramdisk image
	new file:   recipes-bsp/grub/files/40_custom     - grub bootloader options
	new file:   recipes-bsp/grub/grub_1.99.bbappend
	new file:   recipes-core/base-files/base-files_3.0.14.bbappend
	new file:   recipes-core/base-files/files/fstab  - updated fstab per new partition table
	new file:   recipes-core/initramfs/files/init-boot.sh  - ramdisk init script to implement boot modes
	new file:   recipes-core/initramfs/files/init-install.sh - installer script
	new file:   recipes-core/initramfs/initramfs-check-boot_1.0.bb
	new file:   recipes-core/initramfs/initramfs-check-live-install_1.0.bb